test(vdom): extract shared dj-if + sync_ids test harness (#1415)
Pure refactor. After #1408 + #1412, the same VDOM test helpers were
duplicated across three integration test files:

- tests/test_dj_if_keyed_diff_1358.rs (#1358)
- tests/test_sync_ids_dj_if_1408.rs   (#1408)
- tests/torture_round_trip_with_sync.rs (#1412)

Extracts the shared bits to `crates/djust_vdom/tests/common/mod.rs`
(directory form, so cargo doesn't compile it as its own test target):

  - dj-if marker helpers: dj_if_open, dj_if_close,
    is_dj_if_open / is_dj_if_open_any / is_dj_if_close,
    match_close_idx
  - VNode lookup: find_by_djust_id, find_by_djust_id_mut
  - dj-if subtree manipulation: find_dj_if_subtree,
    remove_dj_if_subtree
  - apply_all (subtree-aware patch applier — handles
    Insert/RemoveSubtree, which the crate's apply_patches skips)
  - assert_handles_resolve (cross-render invariant checker)
  - IdGen (sequential base62 dj-id generator)
  - elem / elem_with_text (id-stamped VNode constructors)
  - count (patch-count helper)

Test files now `mod common;` and import what they need. The
`Cell<u64>` counter pattern is replaced with `IdGen` to give the
shared helpers a clean type.

Test count unchanged (232 pass). Unblocks #1413 (proptest), #1416
(HTML round-trip), #1417 (ignore × dj-if × sync_ids), #1418 (deep
cascade), #1419 (wire-protocol snapshot), #1420 (patch-batch
ordering) — all of which need the same helpers.
